+++ changelog.md
@@ -4,6 +4,18 @@
 
 ## Unreleased
 
+### Fixed
+
+- Fixes the check for a non-existent flag (issue #40)
+
+---
+
+## 0.3.0a1 - 2019-10-09
+
+### Fixed
+
+- Fixes the check for a non-existent flag (issue #40)
+
 ---
 
 ## 0.3.0a0 - 2019-10-07